LPS-28479 Source formatting

+++ portal-impl/src/com/liferay/portal/velocity/VelocityEngineImpl.java
@@ -57,7 +57,9 @@ public class VelocityEngineImpl implements VelocityEngine {
 		StringResourceRepository stringResourceRepository =
 			StringResourceLoader.getRepository();
 
-		stringResourceRepository.removeStringResource(velocityTemplateId);
+		if (stringResourceRepository != null) {
+			stringResourceRepository.removeStringResource(velocityTemplateId);
+		}
 
 		LiferayResourceCacheUtil.remove(
 			_getResourceCacheKey(velocityTemplateId));
